src/api_server/requirements.txt
fastapi
uvicorn
kafka-python
pydantic-settings

# Database
sqlalchemy
psycopg2-binary